** Refactoring commit: In Common project the package logs renamed to logger.
