+++ Team6.Server/src/ApplicationEntryPoint.java
@@ -11,7 +11,7 @@ import javafx.fxml.FXMLLoader;
 import javafx.scene.Parent;
 import javafx.scene.Scene;
 import javafx.stage.Stage;
-import logs.LogManager;
+import logger.LogManager;
 import common.StartableState;
 import common.UncaughetExceptions;
