+++ Team6.Server/src/ServerController.java
@@ -34,8 +34,8 @@ import javafx.scene.control.TextInputDialog;
 import javafx.scene.control.cell.PropertyValueFactory;
 import javafx.scene.paint.Paint;
 import javafx.scene.shape.Circle;
-import logs.LogFormatter;
-import logs.LogManager;
+import logger.LogFormatter;
+import logger.LogManager;
 import messages.EntityData;
 import messages.EntityDataOperation;
 import messages.Message;
